System.IO.Pipes
