+++ src/System.IO.Pipes/tests/NamedPipeTests/NamedPipeTest.cs
