+++ src/System.IO.Pipes/tests/PipeTest.cs
